separate container and template for organizer block
