+++ src/modules/blocks/event-organizer/index.js
@@ -1,8 +1,4 @@
 /**
- * External dependencies
- */
-
-/**
  * WordPress dependencies
  */
 import { __ } from '@wordpress/i18n';
@@ -11,10 +7,8 @@ import { __ } from '@wordpress/i18n';
  *
  * Internal dependencies
  */
-import Organizer from './block';
+import Organizer from './container';
 import { Icons } from 'elements';
-import { withStore } from 'editor/hoc';
-import { ORGANIZER } from 'editor/post-types';
 
 export default {
 	id: 'event-organizer',
@@ -40,7 +34,7 @@ export default {
 		},
 	},
 
-	edit: withStore( { isolated: true, postType: ORGANIZER } )( Organizer ),
+	edit: Organizer,
 
 	save() {
 		return null;
