+++ src/modules/blocks/event-organizer/template.js
@@ -4,8 +4,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { isEmpty } from 'lodash';
-import { compose, bindActionCreators } from 'redux';
-import { connect } from 'react-redux';
 
 /**
  * WordPress dependencies
@@ -21,7 +19,7 @@ import {
 } from '@wordpress/editor';
 
 /**
- * Internal dependendencies
+ * Internal dependencies
  */
 import {
 	SearchOrCreate,
@@ -29,10 +27,7 @@ import {
 } from 'elements';
 import OrganizerDetails from './details';
 import OrganizerForm from './details/form';
-import { withSaveData, withDetails, withForm } from 'editor/hoc';
 import OrganizerIcon from 'icons/organizer.svg';
-import { actions, selectors } from 'data/blocks/organizers';
-import { actions as detailsActions } from 'data/details';
 import { toFields, toOrganizer } from 'elements/organizer-form/utils';
 
 class Organizer extends Component {
@@ -51,10 +46,6 @@ class Organizer extends Component {
 		createDraft: PropTypes.func,
 		editPost: PropTypes.func,
 	};
-
-	constructor( props ) {
-		super( ...arguments );
-	}
 
 	componentDidUpdate( prevProps ) {
 		const {
@@ -235,34 +226,4 @@ class Organizer extends Component {
 	};
 }
 
-const mapStateToProps = ( state, props ) => {
-	return {
-		organizer: selectors.getOrganizerInBlock( state, props ),
-		organizers: selectors.getOrganizersInClassic( state ),
-	};
-};
-
-const mapDispatchToProps = ( dispatch ) => {
-	return {
-		...bindActionCreators( actions, dispatch ),
-		...bindActionCreators( detailsActions, dispatch ),
-		setInitialState( { id, get } ) {
-			const organizer = get( 'organizer', '' );
-			if ( ! organizer ) {
-				return;
-			}
-			dispatch( actions.addOrganizerInBlock( id, organizer ) );
-			dispatch( actions.addOrganizerInClassic( organizer ) );
-		},
-	};
-};
-
-export default compose(
-	connect(
-		mapStateToProps,
-		mapDispatchToProps,
-	),
-	withDetails( 'organizer' ),
-	withSaveData(),
-	withForm( ( props ) => props.id ),
-)( Organizer );
+export default Organizer;
